Update PrismData.py
Had to change the Bio.SubsMat to Align, though this might interfere with some of the actual use of PrismData, but those functions aren't gonna be used in the current project. So changing the import to Align.substitution_matrices lets the google colab run smoothly since there then isn't any issue for the helpers.py file to import other functions from PrismData.